+++ apps/admin/pages/live-shopping/[liveShoppingId].tsx
@@ -1,168 +1,33 @@
-import { ChevronLeftIcon, EditIcon } from '@chakra-ui/icons';
-import {
-  Accordion,
-  AccordionButton,
-  AccordionIcon,
-  AccordionItem,
-  AccordionPanel,
-  Alert,
-  AlertDescription,
-  Box,
-  Button,
-  Divider,
-  Flex,
-  Grid,
-  Input,
-  Link,
-  Radio,
-  RadioGroup,
-  Stack,
-  Text,
-  Textarea,
-  useDisclosure,
-  useToast,
-} from '@chakra-ui/react';
-import { GoodsConfirmationStatuses, TtsSetting } from '@prisma/client';
-import { AdminLiveShoppingBroadcasterName } from '@project-lc/components-admin/AdminLiveShoppingBroadcasterName';
-import { AdminLiveShoppingUpdateConfirmModal } from '@project-lc/components-admin/AdminLiveShoppingUpdateConfirmModal';
-import { AdminOverlayImageUploadDialog } from '@project-lc/components-admin/AdminOverlayImageUploadDialog';
+import { ChevronLeftIcon } from '@chakra-ui/icons';
+import { Box, Button, Flex, Grid, Stack } from '@chakra-ui/react';
 import { AdminPageLayout } from '@project-lc/components-admin/AdminPageLayout';
-import { BroadcasterAutocomplete } from '@project-lc/components-admin/BroadcasterAutocomplete';
-import { LiveShoppingDatePicker } from '@project-lc/components-admin/LiveShoppingDatePicker';
+import { AdminLiveShoppingUpdateForm } from '@project-lc/components-admin/live-shopping/AdminLiveShoppingUpdateForm';
+import { LiveShoppingCurrentDataDisplay } from '@project-lc/components-admin/live-shopping/LiveShoppingCurrentDataDisplay';
+import { LiveShoppingKkshowGoodsDetailDisplay } from '@project-lc/components-admin/live-shopping/LiveShoppingKkshowGoodsDetailDisplay';
 import { LiveShoppingDetailTitle } from '@project-lc/components-admin/LiveShoppingDetailTitle';
-import { AdminLiveShoppingSpecialPrice } from '@project-lc/components-admin/live-shopping/AdminLiveShoppingSpecialPrice';
-import { LiveShoppingProgressSelector } from '@project-lc/components-admin/LiveShoppingProgressSelector';
-import { SectionWithTitle } from '@project-lc/components-layout/SectionWithTitle';
-import { GoodsDetailCommonInfo } from '@project-lc/components-seller/goods-detail/GoodsDetailCommonInfo';
-import { GoodsDetailImagesInfo } from '@project-lc/components-seller/goods-detail/GoodsDetailImagesInfo';
-import { GoodsDetailInfo } from '@project-lc/components-seller/goods-detail/GoodsDetailInfo';
-import { GoodsDetailOptionsInfo } from '@project-lc/components-seller/goods-detail/GoodsDetailOptionsInfo';
-import { GoodsDetailPurchaseLimitInfo } from '@project-lc/components-seller/goods-detail/GoodsDetailPurchaseLimitInfo';
-import { GoodsDetailShippingInfo } from '@project-lc/components-seller/goods-detail/GoodsDetailShippingInfo';
-import { GoodsDetailSummary } from '@project-lc/components-seller/goods-detail/GoodsDetailSummary';
-import { LiveShoppingProgressBadge } from '@project-lc/components-shared/LiveShoppingProgressBadge';
-import {
-  LiveShoppingManage,
-  useAdminGoodsById,
-  useAdminLiveShoppingList,
-  useProfile,
-  useUpdateLiveShoppingManageMutation,
-} from '@project-lc/hooks';
-import { LiveShoppingUpdateDTO, LIVE_SHOPPING_PROGRESS } from '@project-lc/shared-types';
-import { AxiosError } from 'axios';
-import dayjs from 'dayjs';
+import { useAdminLiveShoppingList, useProfile } from '@project-lc/hooks';
 import { useRouter } from 'next/router';
-import { useState } from 'react';
-import { FormProvider, useForm } from 'react-hook-form';
 
-function getDuration(startDate: Date, endDate: Date): string {
-  if (startDate && startDate) {
-    const startTime = dayjs(startDate);
-    const endTime = dayjs(endDate);
-    const duration = endTime.diff(startTime);
-    const hours = duration / (1000 * 60 * 60);
-    const stringfiedHours = hours.toFixed(1);
-    return `${stringfiedHours}시간`;
-  }
-  return '미정';
-}
-
-export type LiveShoppingFormData = Omit<LiveShoppingUpdateDTO, 'id'>;
 export function LiveShoppingDetail(): JSX.Element {
   const router = useRouter();
   const liveShoppingId = router.query.liveShoppingId as string;
+
   const { data: profileData } = useProfile();
-  const {
-    data: liveShopping,
-    isLoading: liveShoppingIsLoading,
-    refetch,
-  } = useAdminLiveShoppingList(
-    { id: Number(liveShoppingId) },
-    {
-      enabled:
-        !!profileData?.id || !liveShoppingId || !Number.isNaN(Number(liveShoppingId)),
-    },
-  );
-
-  const goodsId = liveShopping ? liveShopping[0].goodsId : '';
-  const goods = useAdminGoodsById(goodsId);
-
-  const { mutateAsync } = useUpdateLiveShoppingManageMutation();
-  const methods = useForm<LiveShoppingFormData>({
-    defaultValues: {
-      progress: undefined,
-      liveShoppingName: '',
-      broadcasterId: undefined,
-      broadcastStartDate: '',
-      broadcastEndDate: '',
-      sellStartDate: '',
-      sellEndDate: '',
-      rejectionReason: '',
-      videoUrl: '',
-      whiletrueCommissionRate: undefined,
-      broadcasterCommissionRate: undefined,
-    },
-  });
-  const toast = useToast();
-  const [isOpen, setIsOpen] = useState(false);
-  const {
-    isOpen: imageDialogIsOpen,
-    onOpen: imageDialogOnOpen,
-    onClose: imageDialogOnClose,
-  } = useDisclosure();
-  const onClose = (): void => {
-    setIsOpen(false);
-  };
-
-  const openConfirmModal = (): void => {
-    setIsOpen(true);
-  };
-
-  const onSuccess = (): void => {
-    reset({
-      progress: undefined,
-      liveShoppingName: '',
-      broadcasterId: undefined,
-      broadcastStartDate: '',
-      broadcastEndDate: '',
-      sellStartDate: '',
-      sellEndDate: '',
-      rejectionReason: '',
-      videoUrl: '',
-      whiletrueCommissionRate: undefined,
-      broadcasterCommissionRate: undefined,
-      messageSetting: {
-        fanNick: '',
-        levelCutOffPoint: undefined,
-        ttsSetting: undefined,
+  const { data: liveShoppings, isLoading: liveShoppingIsLoading } =
+    useAdminLiveShoppingList(
+      { id: Number(liveShoppingId) },
+      {
+        enabled:
+          !!profileData?.id && !!liveShoppingId && !Number.isNaN(Number(liveShoppingId)),
       },
-    });
-    toast({ title: '변경 완료', status: 'success' });
-    refetch();
-  };
+    );
 
-  const onFail = (err?: AxiosError): void => {
-    toast({
-      title: '변경 실패',
-      description: err.response.status === 400 ? err?.response?.data?.message : undefined,
-      status: 'error',
-    });
-  };
+  if (liveShoppingIsLoading) return <AdminPageLayout>...loading</AdminPageLayout>;
+  if (!liveShoppings) return <AdminPageLayout>...no live-shoppings</AdminPageLayout>;
 
-  const { handleSubmit, register, watch, reset } = methods;
-  const onSubmit = async (data: LiveShoppingFormData): Promise<void> => {
-    const videoUrlExist = Boolean(liveShopping[0]?.liveShoppingVideo?.youtubeUrl);
-    const { sellerId, goodsId: _, contactId, requests, ...restData } = data; // formData타입에서 LiveShoppingManage 타입으로 바꾸기 위해
-    const dto: LiveShoppingManage = Object.assign(restData, {
-      id: Number(liveShoppingId),
-    });
-    mutateAsync({ dto, videoUrlExist }).then(onSuccess).catch(onFail);
-  };
-  if (liveShoppingIsLoading || goods.isLoading)
-    return <AdminPageLayout>...loading</AdminPageLayout>;
-
-  if (!goods.isLoading && !goods.data)
-    return <AdminPageLayout>...no data</AdminPageLayout>;
+  const currentLiveShopping = liveShoppings[0];
+  if (!currentLiveShopping)
+    return <AdminPageLayout>...no live-shopping data</AdminPageLayout>;
 
   return (
     <AdminPageLayout>
@@ -179,414 +44,23 @@ export function LiveShoppingDetail(): JSX.Element {
           </Flex>
         </Box>
         {/* 상품 제목 */}
-        {liveShopping && !liveShoppingIsLoading && (
+        {liveShoppings && !liveShoppingIsLoading && (
           <LiveShoppingDetailTitle
-            liveShoppingName={liveShopping[0].liveShoppingName}
-            createDate={liveShopping[0].createDate}
+            liveShoppingName={currentLiveShopping.liveShoppingName}
+            createDate={currentLiveShopping.createDate}
           />
         )}
         <Grid templateColumns="repeat(2, 1fr)" justifyItems="start" gap={4}>
-          <Stack spacing={5}>
-            <Box>
-              <Stack direction="row">
-                <Text as="span">상품명 :</Text>
-                <Text color="blue">
-                  <Text as="span" color="red">
-                    {liveShopping[0].goods.confirmation?.status ===
-                      GoodsConfirmationStatuses.waiting ||
-                      (liveShopping[0].goods.confirmation?.status ===
-                        GoodsConfirmationStatuses.needReconfirmation &&
-                        `(검수미완료) `)}
-                    {liveShopping[0].goods.confirmation?.status ===
-                      GoodsConfirmationStatuses.rejected && `(검수거절상품) `}
-                  </Text>
-                  {liveShopping[0].broadcaster
-                    ? `${liveShopping[0].goods.goods_name} + ${liveShopping[0].broadcaster.userNickname}`
-                    : `${liveShopping[0].goods.goods_name}`}
-                </Text>
-              </Stack>
-              {liveShopping[0].goods.confirmation?.status !==
-                GoodsConfirmationStatuses.confirmed && (
-                <Alert status="info" variant="left-accent" rounded="md">
-                  <AlertDescription>
-                    {liveShopping[0].goods.confirmation?.status ===
-                    GoodsConfirmationStatuses.rejected
-                      ? '해당 상품은 검수 거절된 상품입니다. (라이브쇼핑을 취소하세요.)'
-                      : '검수 미완료 상태이므로 상품 검수부터 진행해야 합니다.'}
-                    <Button
-                      size="xs"
-                      onClick={() => router.push(`/goods/${liveShopping[0].goodsId}`)}
-                    >
-                      상품 검수 이동하기
-                    </Button>
-                  </AlertDescription>
-                </Alert>
-              )}
-            </Box>
-            <Stack direction="row">
-              <Text as="span">판매자 :</Text>
-              <Text color="blue">
-                {liveShopping[0].seller.sellerShop?.shopName || ''}
-              </Text>
-            </Stack>
-
-            <Stack direction="row" alignItems="center">
-              <Text as="span">현재 진행상태</Text>
-              <LiveShoppingProgressBadge
-                progress={liveShopping[0].progress}
-                broadcastStartDate={liveShopping[0].broadcastStartDate}
-                broadcastEndDate={liveShopping[0].broadcastEndDate}
-                sellEndDate={liveShopping[0].sellEndDate}
-              />
-              {liveShopping[0].progress === LIVE_SHOPPING_PROGRESS.취소됨 ? (
-                <Text>사유 : {liveShopping[0].rejectionReason}</Text>
-              ) : null}
-            </Stack>
-
-            <Divider />
-            <Stack direction="row" alignItems="center">
-              <Text as="span">방송인: </Text>
-              {liveShopping[0].broadcaster ? (
-                <AdminLiveShoppingBroadcasterName
-                  data={liveShopping[0].broadcaster}
-                  color="blue"
-                />
-              ) : (
-                <Text fontWeight="bold">미정</Text>
-              )}
-            </Stack>
-            <Stack direction="row" alignItems="center">
-              <Text as="span">방송시작 시간: </Text>
-              <Text as="span" fontWeight="bold" color="blue">
-                {liveShopping[0].broadcastStartDate
-                  ? dayjs(liveShopping[0].broadcastStartDate).format('YYYY/MM/DD HH:mm')
-                  : '미정'}
-              </Text>
-            </Stack>
-
-            <Stack direction="row" alignItems="center">
-              <Text as="span">방송종료 시간: </Text>
-              <Text as="span" fontWeight="bold" color="blue">
-                {liveShopping[0].broadcastEndDate
-                  ? dayjs(liveShopping[0].broadcastEndDate).format('YYYY/MM/DD HH:mm')
-                  : '미정'}
-              </Text>
-            </Stack>
-
-            <Stack direction="row" alignItems="center">
-              <Text as="span">방송시간: </Text>
-              <Text as="span" fontWeight="bold" color="blue">
-                {getDuration(
-                  liveShopping[0].broadcastStartDate,
-                  liveShopping[0].broadcastEndDate,
-                )}
-              </Text>
-            </Stack>
-            {liveShopping[0].progress === 'confirmed' &&
-              liveShopping[0].liveShoppingVideo && (
-                <Stack direction="row" alignItems="center">
-                  <Text as="span">영상 URL: </Text>
-                  <Text as="span" fontWeight="bold" color="blue">
-                    <Link
-                      isTruncated
-                      href={liveShopping[0].liveShoppingVideo.youtubeUrl || ''}
-                      fontWeight="bold"
-                      colorScheme="blue"
-                      textDecoration="underline"
-                      isExternal
-                    >
-                      {liveShopping[0].liveShoppingVideo.youtubeUrl || ''}
-                    </Link>
-                  </Text>
-                </Stack>
-              )}
-
-            <Divider />
-
-            <Stack direction="row" alignItems="center">
-              <Text as="span">판매시작 시간: </Text>
-              <Text as="span" fontWeight="bold" color="blue">
-                {liveShopping[0].sellStartDate
-                  ? dayjs(liveShopping[0].sellStartDate).format('YYYY/MM/DD HH:mm')
-                  : '미정'}
-              </Text>
-            </Stack>
-
-            <Stack direction="row" alignItems="center">
-              <Text as="span">판매종료 시간: </Text>
-              <Text as="span" fontWeight="bold" color="blue">
-                {liveShopping[0].sellEndDate
-                  ? dayjs(liveShopping[0].sellEndDate).format('YYYY/MM/DD HH:mm')
-                  : '미정'}
-              </Text>
-            </Stack>
-            <Stack direction="row" alignItems="center">
-              <Text as="span">판매시간: </Text>
-              <Text as="span" fontWeight="bold" color="blue">
-                {getDuration(liveShopping[0].sellStartDate, liveShopping[0].sellEndDate)}
-              </Text>
-            </Stack>
-
-            <Box>
-              <Text fontWeight="bold">구매 메시지 설정</Text>
-              <Text>비회원 팬닉: </Text>
-              <Text as="span" fontWeight="bold" color="blue">
-                {liveShopping[0].messageSetting?.fanNick || ''}
-              </Text>
-
-              <Text>1,2단계 구매메시지 기준 금액: </Text>
-              <Text as="span" fontWeight="bold" color="blue">
-                {liveShopping[0].messageSetting?.levelCutOffPoint || ''}
-              </Text>
-
-              <Text>TTS 설정: </Text>
-              <Text as="span" fontWeight="bold" color="blue">
-                {liveShopping[0].messageSetting?.ttsSetting || ''}
-              </Text>
-            </Box>
-
-            <Divider />
-            <Stack direction="row" alignItems="center">
-              <Text as="span">희망 판매 수수료: </Text>
-              <Text as="span" fontWeight="bold" color="blue">
-                {liveShopping[0].desiredCommission} %
-              </Text>
-            </Stack>
-
-            <Stack direction="row" alignItems="center">
-              <Text as="span">희망 진행 기간: </Text>
-              <Text as="span" fontWeight="bold" color="blue">
-                {liveShopping[0].desiredPeriod}
-              </Text>
-            </Stack>
-
-            <Divider />
-            <Stack direction="row" alignItems="center">
-              <Text as="span">방송인 수수료: </Text>
-              <Text as="span" fontWeight="bold" color="blue">
-                {liveShopping[0].broadcasterCommissionRate
-                  ? `${liveShopping[0].broadcasterCommissionRate}%`
-                  : '미정'}
-              </Text>
-            </Stack>
-
-            <Stack direction="row" alignItems="center">
-              <Text as="span">와일트루 수수료: </Text>
-              <Text as="span" fontWeight="bold" color="blue">
-                {liveShopping[0].whiletrueCommissionRate
-                  ? `${liveShopping[0].whiletrueCommissionRate}%`
-                  : '미정'}
-              </Text>
-            </Stack>
-
-            <Divider />
-
-            {/* 라이브쇼핑 특가정보 표시 및 수정 */}
-            <AdminLiveShoppingSpecialPrice liveShopping={liveShopping[0]} />
-
-            <Divider />
-
-            <Box>
-              <Text>요청사항</Text>
-              <Textarea
-                resize="none"
-                rows={10}
-                value={liveShopping[0].requests || ''}
-                readOnly
-              />
-            </Box>
-          </Stack>
-
+          {/* 현재 라이브쇼핑 데이터 표시 */}
+          <LiveShoppingCurrentDataDisplay liveShopping={currentLiveShopping} />
           {/* 라이브쇼핑 정보 변경 폼 */}
-          <FormProvider {...methods}>
-            <Stack as="form" spacing={5}>
-              <LiveShoppingProgressSelector />
-              <Divider />
-              <Stack>
-                <Text>라이브 쇼핑 이름</Text>
-                <Input {...register('liveShoppingName')} />
-              </Stack>
-              <BroadcasterAutocomplete />
-
-              <Divider />
-              <Box>
-                <Text fontWeight="bold">구매 메시지 설정</Text>
-                <Text>비회원 팬닉: </Text>
-                <Input {...register('messageSetting.fanNick')} placeholder="OOO 팬" />
-
-                <Text>1,2단계 구매메시지 기준 금액: </Text>
-                <Input
-                  type="number"
-                  {...register('messageSetting.levelCutOffPoint', {
-                    valueAsNumber: true,
-                  })}
-                  placeholder="30000"
-                />
-
-                <Text>TTS 설정: </Text>
-                <RadioGroup
-                  onChange={(value: TtsSetting) => {
-                    methods.setValue('messageSetting.ttsSetting', value);
-                  }}
-                >
-                  <Stack spacing={2} direction="row" flexWrap="wrap">
-                    <Radio
-                      {...register('messageSetting.ttsSetting')}
-                      value={TtsSetting.full}
-                    >
-                      전체
-                    </Radio>
-                    <Radio
-                      {...register('messageSetting.ttsSetting')}
-                      value={TtsSetting.nick_purchase}
-                    >
-                      닉,구매감사
-                    </Radio>
-                    <Radio
-                      {...register('messageSetting.ttsSetting')}
-                      value={TtsSetting.nick_purchase_price}
-                    >
-                      닉,금액,구매감사
-                    </Radio>
-                    <Radio
-                      {...register('messageSetting.ttsSetting')}
-                      value={TtsSetting.only_message}
-                    >
-                      메세지만
-                    </Radio>
-                    <Radio
-                      {...register('messageSetting.ttsSetting')}
-                      value={TtsSetting.no_tts}
-                    >
-                      TTS 없음
-                    </Radio>
-                    <Radio
-                      {...register('messageSetting.ttsSetting')}
-                      value={TtsSetting.no_sound}
-                    >
-                      소리 없음(콤보모드)
-                    </Radio>
-                  </Stack>
-                </RadioGroup>
-              </Box>
-
-              <Divider />
-
-              <LiveShoppingDatePicker
-                title="방송 시작시간"
-                registerName="broadcastStartDate"
-              />
-              <LiveShoppingDatePicker
-                title="방송 종료시간"
-                registerName="broadcastEndDate"
-              />
-              {dayjs(watch('broadcastStartDate')) > dayjs(watch('broadcastEndDate')) && (
-                <Text as="em" color="tomato">
-                  방송 종료시간이 시작시간보다 빠릅니다
-                </Text>
-              )}
-              <Divider />
-
-              <LiveShoppingDatePicker
-                title="판매 시작시간"
-                registerName="sellStartDate"
-              />
-              <LiveShoppingDatePicker
-                title="판매 종료시간"
-                registerName="sellEndDate"
-                min={watch('sellStartDate')}
-              />
-              {dayjs(watch('sellStartDate')) > dayjs(watch('sellEndDate')) && (
-                <Text as="em" color="tomato">
-                  판매 종료시간이 시작시간보다 빠릅니다
-                </Text>
-              )}
-              <Divider />
-
-              <Stack>
-                <Text>
-                  영상 URL ( https://youtu.be/4pIuCJTMXQU 와 같은 형태로 입력해주세요 )
-                </Text>
-                <Input
-                  placeholder="https://youtu.be/4pIuCJTMXQU"
-                  {...register('videoUrl')}
-                />
-              </Stack>
-              <Divider />
-
-              <Button onClick={openConfirmModal} colorScheme="blue">
-                변경
-              </Button>
-              <Button
-                rightIcon={<EditIcon />}
-                onClick={imageDialogOnOpen}
-                isDisabled={!liveShopping[0].broadcaster}
-              >
-                오버레이 이미지 등록
-              </Button>
-            </Stack>
-
-            <AdminLiveShoppingUpdateConfirmModal
-              isOpen={isOpen}
-              onClose={onClose}
-              onConfirm={handleSubmit(onSubmit)}
-            />
-            {liveShopping[0].broadcaster && (
-              <AdminOverlayImageUploadDialog
-                isOpen={imageDialogIsOpen}
-                onClose={imageDialogOnClose}
-                broadcasterEmail={liveShopping[0].broadcaster.email}
-                liveShoppingId={liveShopping[0].id}
-              />
-            )}
-          </FormProvider>
+          <AdminLiveShoppingUpdateForm liveShopping={currentLiveShopping} />
         </Grid>
 
-        <Accordion allowMultiple>
-          <AccordionItem>
-            <AccordionButton>
-              <Box flex="1" textAlign="left">
-                상품 정보 보기
-              </Box>
-              <AccordionIcon />
-            </AccordionButton>
-
-            <AccordionPanel pb={4}>
-              {/* 상품 요약 */}
-              <Box as="section">
-                <GoodsDetailSummary goods={goods.data} />
-              </Box>
-
-              {/* 상품 정보 */}
-
-              <SectionWithTitle title="기본 정보">
-                <GoodsDetailInfo goods={goods.data} />
-              </SectionWithTitle>
-
-              <SectionWithTitle title="상품사진 및 설명">
-                <GoodsDetailImagesInfo goods={goods.data} />
-              </SectionWithTitle>
-
-              <SectionWithTitle title="옵션">
-                <GoodsDetailOptionsInfo goods={goods.data} />
-              </SectionWithTitle>
-
-              <SectionWithTitle title="상품 공통 정보">
-                <GoodsDetailCommonInfo goods={goods.data} />
-              </SectionWithTitle>
-
-              <SectionWithTitle title="구매 제한">
-                <GoodsDetailPurchaseLimitInfo goods={goods.data} />
-              </SectionWithTitle>
-
-              {goods.data.ShippingGroup && (
-                <SectionWithTitle title="배송정책">
-                  <GoodsDetailShippingInfo goods={goods.data} />
-                </SectionWithTitle>
-              )}
-            </AccordionPanel>
-          </AccordionItem>
-        </Accordion>
+        {/* 크크쇼 상품으로 라이브 진행하는 경우 크크쇼 상품 상세정보 표시 */}
+        {currentLiveShopping.goodsId && (
+          <LiveShoppingKkshowGoodsDetailDisplay goodsId={currentLiveShopping.goodsId} />
+        )}
       </Stack>
     </AdminPageLayout>
   );
